Remove state.language from dependency array, fixing bug where after switching languages, multiple study popups would render
